// zichange-api-master/src/kyc/sumsub/types/sumsub.supported-document-types.ts
export enum SumsubSupportedDocumentTypes {
  ID_CARD = 'ID_CARD',
  PASSPORT = 'PASSPORT',
  DRIVERS = 'DRIVERS',
  BANK_CARD = 'BANK_CARD',
  UTILITY_BILL = 'UTILITY_BILL',
  SNILS = 'SNILS',
  SELFIE = 'SELFIE',
  PROFILE_IMAGE = 'PROFILE_IMAGE',
  ID_DOC_PHOTO = 'ID_DOC_PHOTO',
  AGREEMENT = 'AGREEMENT',
  CONTRACT = 'CONTRACT',
  RESIDENCE_PERMIT = 'RESIDENCE_PERMIT',
  EMPLOYMENT_CERTIFICATE = 'EMPLOYMENT_CERTIFICATE',
  DRIVERS_TRANSLATION = 'DRIVERS_TRANSLATION',
  INVESTOR_DOC = 'INVESTOR_DOC',
  VEHICLE_REGISTRATION_CERTIFICATE = 'VEHICLE_REGISTRATION_CERTIFICATE',
  INCOME_SOURCE = 'INCOME_SOURCE',
  OTHER = 'OTHER',
}